avoid compile warnings

+++ android/app/src/main/java/de/wellenvogel/ochartsprovider/Settings.java
@@ -44,7 +44,7 @@ public class Settings {
         return testMode;
     }
 
-    public int getShutdownSec() {
+    public long getShutdownSec() {
         return shutdownSec;
     }
